+++ lite2/provider/http/http.go
@@ -13,6 +13,8 @@ import (
 type SignStatusClient interface {
 	rpcclient.SignClient
 	rpcclient.StatusClient
+	// Remote returns the remote network address in a string form.
+	Remote() string
 }
 
 // http provider uses an RPC client (or SignStatusClient more generally) to
@@ -43,6 +45,10 @@ func NewWithClient(chainID string, client SignStatusClient) provider.Provider {
 // ChainID returns a chainID this provider was configured with.
 func (p *http) ChainID() string {
 	return p.chainID
+}
+
+func (p *http) String() string {
+	return fmt.Sprintf("http{%s}", p.client.Remote())
 }
 
 // SignedHeader fetches a SignedHeader at the given height and checks the
