+++ lite2/provider/provider.go
@@ -1,6 +1,8 @@
 package provider
 
-import "github.com/tendermint/tendermint/types"
+import (
+	"github.com/tendermint/tendermint/types"
+)
 
 // Provider provides information for the lite client to sync (verification
 // happens in the client).
